Exploit new blocking outbox in StreamKafkaP
Also add cancellation check to test and rename to new convention
